Separated UserBannerService class
