+++ htdocs/admin/content/user/userbanner.inc.php
@@ -1,6 +1,6 @@
 <?PHP
 
-use EtoA\Ranking\RankingService;
+use EtoA\Ranking\UserBannerService;
 use EtoA\User\UserRepository;
 
 echo "<h1>Spieler-Banner</h1>";
@@ -10,12 +10,12 @@ echo "<p>Banner werden jeweils beim Aktualisieren der Punkte neu generiert.</p>"
 /** @var UserRepository $userRepository */
 $userRepository = $app[UserRepository::class];
 
-/** @var RankingService $rankingService */
-$rankingService = $app[RankingService::class];
+/** @var UserBannerService $userBannerService */
+$userBannerService = $app[UserBannerService::class];
 
 $userNicks = $userRepository->searchUserNicknames();
 foreach ($userNicks as $userId => $userNick) {
-    $name = $rankingService->getUserBannerPath($userId);
+    $name = $userBannerService->getUserBannerPath($userId);
     if (file_exists($name)) {
         echo '<img src="' . $name . '" alt="Banner" style="width:' . USERBANNER_WIDTH . 'px;heigth:' . USERBANNER_HEIGTH . 'px;" /> ';
     } else {
