+++ htdocs/content/userconfig/banner.php
@@ -1,13 +1,13 @@
 <?PHP
 
 use EtoA\Core\Configuration\ConfigurationService;
-use EtoA\Ranking\RankingService;
+use EtoA\Ranking\UserBannerService;
 
 /** @var ConfigurationService */
 $config = $app[ConfigurationService::class];
 
-/** @var RankingService $rankingService */
-$rankingService = $app[RankingService::class];
+/** @var UserBannerService $userBannerService */
+$userBannerService = $app[UserBannerService::class];
 
 $id = $cu->id;
 
@@ -15,7 +15,7 @@ iBoxStart("Banner");
 echo 'Hilf mit, EtoA bekannter zu machen und binde unser Banner auf deiner Website ein!
 Hier findest du den Quellcode um das Banner einzubinden:<br><br>';
 
-$name = $rankingService->getUserBannerPath($id);
+$name = $userBannerService->getUserBannerPath($id);
 if (file_exists($name)) {
     echo '<div style="text-align: center;">
     <img src="' . $name . '" alt="Banner"><br><br>
